moved the code from app to nav bar
